requirements.txt
rasa==2.8.25
python-decouple==3.6
sanic~=22.3.2
watchdog~=2.1.7